requirements.txt
matplotlib==3.10.1
nn_torch==0.0.11
numpy==2.2.3
opencv_contrib_python==4.10.0.84
python_mnist==0.7
scipy==1.15.2
torch==2.5.1
